Fix for issue: #64, improve colors for user code expected values (#63)
When testing #55, I noticed that the failure colours for coded tests
looked very bad in dark mode. The issue was that the error colour was
"too light", while the text was also light, making it very hard to read.

This PR changes the extension such that we correctly handle
`.bg-danger`, as well as choosing a slightly more "punchy" colour for
the error colour (which I feel works better between the two themes).