Coded Test Support (#55)
Added support for the new Coded Test feature that will be released with
VectorCAST 24. These changes have no affect on users of older VectorCAST
versions, or users who do not wish to use Coded Tests. The usage of the
Coded Test features is described in the README.md. The following is a
summary of the new features:

- Building an environment with Coded Test support.
- Generating a new Coded Test source file for any unit
- Adding an existing Coded Test source file for any unit
- Editing Coded Test source files, including IntelliSense support
- Update of the test explorer tree to display new tests as they are
added in the editor
- Execution of Coded Tests from the test explorer tree, or by using the
gutter icons in the source editor
- Support for debugging Coded Tests

---------

Co-authored-by: Igor Martinovic <igor.martinovic@vector.com>