Clicast server release (#217)
This PR has all of the changes that make the clicast server mode
"releasable".

The benefit of server mode is much faster interaction with vcast
environments. For example, test execution takes a fraction of a second
vs 2-3 seconds in non-server mode. The time improvement is because
server mode gets rid of the startup cost of vpython and clicast
(initialization, license checkout, environment open, etc.)

Complete usage notes have been added to the README.md, but to summarize:

- A new option has been added to "Use Data Server" --> default ON
- On startup, we check if the VectorCAST version that is configured for
the workspace supports server mode (>vc24sp5)
- If the option is on and vcast supports server mode we start a server
instance, and ping it to verify successful startup
- if startup is successful, we configure the extension to use server
mode, and add a "vDataServer" button to status bar
- The button shows the current server state, and can be used to enable
and disable the server.

---------

Co-authored-by: Denis Moslavac <44575706+Den1552@users.noreply.github.com>
Co-authored-by: SalvatoreZagaria <salvatore.zagaria@vector.com>
Co-authored-by: Andrew V. Teylu <andrew.teylu@vector.com>