Added MCDC Report feature (#243)
## Summary

Added an MC/DC report feature that allows the user to generate a report
for a specific line. The report opens in a separate WebviewPanel,
similar to the test generation report. Lines that are interactable are
visually highlighted in the editor.

## Usage
1) Ensure that `Statement+MCDC` is selected under `Settings → VectorCAST
Test Explorer → Build → Coverage Kind`. If this setting is not selected,
the coverage gutter will display the standard coverage icons and will
not be interactable.

2) To generate a report, `right-click` on the coverage icon and choose
`Get MCDC Report`. The report will open in a separate WebviewPanel.

3) In` Settings → VectorCAST Test Explorer → MCDC Coverage Gutter
Icons`, you can choose to display either an `arrow` or an `M` as the
icon in the coverage gutter.



---------

Co-authored-by: Andrew V. Teylu <andrew.teylu@vector.com>