Requirement Coverage Reports (#46)
Adding a new executable to generate reports (requirements coverage, for now) out of environments where c2r and r2t has been run, based on the code coverage and extracted relevant lines per requirement. We assume code2reqs had been run, but this will need to be generalized to enable the use on projects with existing requirements.

During the implementation I have updated the way we generate custom reports, adapting the algorithm to the reports generated with clicast 25, which are substantially different from the 24 version.

Updated evaluation report. We now have a new general view of the lines that were not covered for each environment and requirement. In each environment's sub-view we also have a list of requirements that didn't cover lines, showing their test cases and linking to the specific coverage report for each requirement.

The new requirements coverage report shows covered lines that had been deemed relevant for a requirement as green, the rest of the code covered by test as light green and the relevant but uncovered lines as red.