Improved identifier fallback (#59)
Gets rid of the need for the test script template by deriving the necessary list of identifiers manually.

This is certainly only approximate however I have gotten good results on a variety of environments (both C and C++). Furthermore this allows us to do interesting things built on top of this in the future (like automatically inferring necessary constructor values etc.)

Potential unimplemented improvements:

Switch to querying the DataAPI instead of reading from params.xml and types.xml
Overloaded functions are not supported right now (but could be by just always supplying the type info for the identifiers at the expense of readability)
Add VECTORCAST_INT and Co. to the identifier list (however we probably do not want the model to touch these by itself for now)

Details:

* Add type resolver to derive VectorCAST identifiers (and other potential usecases)

* Fix: properly support unions

* Track namespaces and construct qualified names for symbols

* Treat a segfault as an error, not a test failure

* Derive allowed identifiers using the type resolver

* Do not sort identifiers

* Update test outputs for test generation (the identifier ordering is slightly different from before)

* Add old ways of getting identifiers as backup

* Add tests for type resolver

* Refactor type resolver

* Simplfy assert (necessary info is available using verbose pytest output)